try composite with iterator Analyzing the composite tree and ways to extend its usage. I've noticed the state pattern as a useful way to configure what the tree does/outputs. For example, if you wanted to: 1. print the tree 2. return the tree as a recursive array 3. recursive array with branch titles
The iterator doesn't appear to function as designed from a polymorphic/composite iteration persective. I'm still using the internal `interpret` method to get back the full tree.

I feel that a dependency injection would be a much better way to inject different kinds of iterators into the composite structure. (Which is analogous to the state pattern)